Issue #3919: [API] Allow to grant user/group permissions on storage paths  (#3923)
(cherry picked from commit 9c2d3680517a14d7d35bd7ec8b47bef6b8743238)

+++ api/src/main/java/com/epam/pipeline/manager/datastorage/DataStorageManager.java
@@ -27,6 +27,7 @@ import com.epam.pipeline.controller.vo.EntityFilterVO;
 import com.epam.pipeline.dao.datastorage.DataStorageDao;
 import com.epam.pipeline.dto.datastorage.lifecycle.restore.StorageRestoreAction;
 import com.epam.pipeline.dto.datastorage.lifecycle.restore.StorageRestorePathType;
+import com.epam.pipeline.dto.datastorage.permissions.StorageFolderListPermissionsContainer;
 import com.epam.pipeline.entity.AbstractSecuredEntity;
 import com.epam.pipeline.entity.BaseEntity;
 import com.epam.pipeline.entity.SecuredEntityWithAction;
@@ -75,6 +76,7 @@ import com.epam.pipeline.entity.utils.DateUtils;
 import com.epam.pipeline.manager.datastorage.lifecycle.DataStorageLifecycleRestoredListingContainer;
 import com.epam.pipeline.manager.datastorage.lifecycle.DataStorageLifecycleManager;
 import com.epam.pipeline.manager.datastorage.lifecycle.DataStorageLifecycleRestoreManager;
+import com.epam.pipeline.manager.datastorage.permissions.StoragePathPermissionsService;
 import com.epam.pipeline.manager.datastorage.providers.ProviderUtils;
 import com.epam.pipeline.manager.metadata.MetadataManager;
 import com.epam.pipeline.manager.docker.ToolVersionManager;
@@ -204,6 +206,9 @@ public class DataStorageManager implements SecuredEntityManager {
 
     @Autowired
     private StoragePermissionManager storagePermissionManager;
+
+    @Autowired
+    private StoragePathPermissionsService storagePathPermissionsService;
 
     private AbstractDataStorageFactory dataStorageFactory =
             AbstractDataStorageFactory.getDefaultDataStorageFactory();
@@ -521,6 +526,7 @@ public class DataStorageManager implements SecuredEntityManager {
         }
         dataStorageLifecycleManager.deleteStorageLifecyclePolicyRules(id);
         storageLifecycleRestoreManager.deleteRestoreActions(id);
+        storagePathPermissionsService.deleteByStorageId(id);
         dataStorageDao.deleteDataStorage(id);
         return dataStorage;
     }
@@ -544,12 +550,13 @@ public class DataStorageManager implements SecuredEntityManager {
         }
         Assert.isTrue(pageSize == null || pageSize > 0,
                 messageHelper.getMessage(MessageConstants.ERROR_PAGE_SIZE));
+        final StorageFolderListPermissionsContainer permissionsContainer = getPermissionsContainer(dataStorage, path);
         if (!showArchived && DataStorageType.S3.equals(dataStorage.getType())) {
             final DataStorageLifecycleRestoredListingContainer restoredListing = loadRestoredPaths(dataStorage, path);
             return storageProviderManager.getRestoredItems(dataStorage, path, showVersion, pageSize, marker,
-                    restoredListing);
+                    restoredListing, permissionsContainer);
         }
-        return storageProviderManager.getItems(dataStorage, path, showVersion, pageSize, marker);
+        return storageProviderManager.getItems(dataStorage, path, showVersion, pageSize, marker, permissionsContainer);
     }
 
     public DataStorageListing filterDataStorageItems(final Long storageId, final String path,
@@ -580,6 +587,7 @@ public class DataStorageManager implements SecuredEntityManager {
             Assert.isTrue(dataStorage.isVersioningEnabled(), messageHelper.getMessage(
                     MessageConstants.ERROR_DATASTORAGE_VERSIONING_REQUIRED, dataStorage.getName()));
         }
+        checkReadPermissionsOnFile(dataStorage, path);
         return storageProviderManager.generateDownloadURL(dataStorage, path, version, contentDisposition);
     }
 
@@ -590,12 +598,15 @@ public class DataStorageManager implements SecuredEntityManager {
         if (paths == null) {
             return urls;
         }
-        paths.forEach(path -> urls.add(storageProviderManager.generateDownloadURL(dataStorage, path, null, null)));
+        paths.stream()
+                .peek(path -> checkReadPermissionsOnFile(dataStorage, path))
+                .forEach(path -> urls.add(storageProviderManager.generateDownloadURL(dataStorage, path, null, null)));
         return urls;
     }
 
     public DataStorageDownloadFileUrl generateDataStorageItemUploadUrl(Long id, String path) {
         AbstractDataStorage dataStorage = load(id);
+        checkWritePermissionsOnFile(dataStorage, path);
         return storageProviderManager.generateDataStorageItemUploadUrl(dataStorage, path);
     }
 
@@ -605,7 +616,9 @@ public class DataStorageManager implements SecuredEntityManager {
         if (paths == null) {
             return urls;
         }
-        paths.forEach(path -> urls.add(storageProviderManager.generateDataStorageItemUploadUrl(dataStorage, path)));
+        paths.stream()
+                .peek(path -> checkWritePermissionsOnFile(dataStorage, path))
+                .forEach(path -> urls.add(storageProviderManager.generateDataStorageItemUploadUrl(dataStorage, path)));
         return urls;
     }
 
@@ -637,6 +650,7 @@ public class DataStorageManager implements SecuredEntityManager {
             byte[] contents)
             throws DataStorageException {
         AbstractDataStorage dataStorage = load(dataStorageId);
+        checkWritePermissionsOnFile(dataStorage, path);
         return storageProviderManager.createFile(dataStorage, path, contents);
     }
 
@@ -647,6 +661,7 @@ public class DataStorageManager implements SecuredEntityManager {
             throws DataStorageException {
         AbstractDataStorage dataStorage = load(dataStorageId);
         String path = getRelativePath(folder, name, dataStorage);
+        checkWritePermissionsOnFile(dataStorage, path);
         return storageProviderManager.createFile(dataStorage, path, contents);
     }
 
@@ -689,6 +704,7 @@ public class DataStorageManager implements SecuredEntityManager {
             List<UpdateDataStorageItemVO> list)
             throws DataStorageException{
         AbstractDataStorage dataStorage = load(dataStorageId);
+        checkPermissionsOnItems(dataStorage, list);
         List<AbstractDataStorageItem> updatedItems = new ArrayList<>();
         for (UpdateDataStorageItemVO item : list) {
             updatedItems.add(updateDataStorageItem(dataStorage, item));
@@ -704,6 +720,7 @@ public class DataStorageManager implements SecuredEntityManager {
             Assert.isTrue(dataStorage.isVersioningEnabled(), messageHelper.getMessage(
                     MessageConstants.ERROR_DATASTORAGE_VERSIONING_REQUIRED, dataStorage.getName()));
         }
+        checkPermissionsOnItems(dataStorage, list);
         for (UpdateDataStorageItemVO item : list) {
             deleteDataStorageItem(dataStorage, item, totally);
         }
@@ -712,6 +729,7 @@ public class DataStorageManager implements SecuredEntityManager {
 
     public DataStorageItemContent getDataStorageItemContent(Long id, String path, String version) {
         AbstractDataStorage dataStorage = load(id);
+        checkReadPermissionsOnFile(dataStorage, path);
         checkDataStorageVersioning(dataStorage, version);
         return storageProviderManager.getFile(dataStorage, path, version);
     }
@@ -719,12 +737,14 @@ public class DataStorageManager implements SecuredEntityManager {
     public Map<String, String> loadDataStorageObjectTags(Long id, String path, String version) {
         AbstractDataStorage dataStorage = load(id);
         checkDataStorageVersioning(dataStorage, version);
+        checkReadPermissionsOnFile(dataStorage, path);
         return storageProviderManager.listObjectTags(dataStorage, path, version);
     }
 
     public Map<String, String> deleteDataStorageObjectTags(Long id, String path, Set<String> tags, String version) {
         AbstractDataStorage dataStorage = load(id);
         checkDataStorageVersioning(dataStorage, version);
+        checkReadPermissionsOnFile(dataStorage, path);
         return storageProviderManager.deleteObjectTags(dataStorage, path, tags, version);
     }
 
@@ -738,6 +758,8 @@ public class DataStorageManager implements SecuredEntityManager {
             return null;
         }
         DataStorageFile dataStorageFile = (DataStorageFile) dataStorageItems.get(0);
+        final AbstractDataStorage dataStorage = load(dataStorageId);
+        checkReadPermissionsOnFile(dataStorage, dataStorageFile.getPath());
         if (MapUtils.isEmpty(dataStorageFile.getVersions())) {
             dataStorageFile.setTags(loadDataStorageObjectTags(dataStorageId, path, null));
         } else {
@@ -752,6 +774,7 @@ public class DataStorageManager implements SecuredEntityManager {
             String version, Boolean rewrite) {
         AbstractDataStorage dataStorage = load(id);
         checkDataStorageVersioning(dataStorage, version);
+        checkReadPermissionsOnFile(dataStorage, path);
         Map<String, String> resultingTags = new HashMap<>();
         if (!rewrite) {
             resultingTags = storageProviderManager.listObjectTags(dataStorage, path, version);
@@ -762,6 +785,7 @@ public class DataStorageManager implements SecuredEntityManager {
 
     public DataStorageStreamingContent getStreamingContent(long dataStorageId, String path, String version) {
         AbstractDataStorage dataStorage = load(dataStorageId);
+        checkReadPermissionsOnFile(dataStorage, path);
         return storageProviderManager.getFileStream(dataStorage, path, version);
     }
 
@@ -1266,5 +1290,72 @@ public class DataStorageManager implements SecuredEntityManager {
                 .filter(action -> StorageRestorePathType.FILE.equals(action.getType()))
                 .map(StorageRestoreAction::getPath)
                 .collect(Collectors.toList());
+    }
+
+    private StorageFolderListPermissionsContainer getPermissionsContainer(final AbstractDataStorage storage,
+                                                                         final String path) {
+        return needToLoadPathPermissions(storage)
+                ? storagePathPermissionsService.getFolderListPermissions(storage.getId(), path)
+                : null;
+    }
+
+    private void checkPermissionsOnItems(final AbstractDataStorage storage, final List<UpdateDataStorageItemVO> items) {
+        if (needToLoadPathPermissions(storage)) {
+            ListUtils.emptyIfNull(items).forEach(item -> checkPermissionsOnItem(storage.getId(), item));
+        }
+    }
+
+    private void checkPermissionsOnItem(final Long storageId, final UpdateDataStorageItemVO item) {
+        switch (item.getType()) {
+            case File:
+                storagePathPermissionsService.canWriteToFile(storageId, item.getPath());
+                if (Objects.isNull(item.getAction())) {
+                    return;
+                }
+                switch (item.getAction()) {
+                    case Move: storagePathPermissionsService.canWriteToFile(storageId, item.getOldPath());
+                    case Copy: storagePathPermissionsService.canReadFile(storageId, item.getOldPath());
+                    default: break;
+                }
+            case Folder:
+                storagePathPermissionsService.canWriteToFolder(storageId, item.getPath());
+                if (Objects.isNull(item.getAction())) {
+                    return;
+                }
+                switch (item.getAction()) {
+                    case Move: storagePathPermissionsService.canWriteToFolder(storageId, item.getOldPath());
+                    case Copy: storagePathPermissionsService.canReadFolder(storageId, item.getOldPath());
+                    default: break;
+                }
+            default: break;
+        }
+    }
+
+    private void checkGetPermissionsOnItem(final AbstractDataStorage storage, final String path,
+                                           final DataStorageItemType itemType) {
+        if (needToLoadPathPermissions(storage)) {
+            switch (itemType) {
+                case Folder: storagePathPermissionsService.canGetFolder(storage.getId(), path);
+                case File: storagePathPermissionsService.canReadFile(storage.getId(), path);
+                default: break;
+            }
+        }
+    }
+
+    private void checkWritePermissionsOnFile(final AbstractDataStorage storage, final String path) {
+        if (needToLoadPathPermissions(storage)) {
+            storagePathPermissionsService.canWriteToFile(storage.getId(), path);
+        }
+    }
+
+    private void checkReadPermissionsOnFile(final AbstractDataStorage storage, final String path) {
+        if (needToLoadPathPermissions(storage)) {
+            storagePathPermissionsService.canReadFile(storage.getId(), path);
+        }
+    }
+
+    private boolean needToLoadPathPermissions(final AbstractDataStorage storage) {
+        return storage.isPathPermissionsEnabled() && DataStorageType.S3.equals(storage.getType())
+                && !authManager.isAdmin() && !authManager.getAuthorizedUser().equalsIgnoreCase(storage.getOwner());
     }
 }
